chatbot_langchain/requirements.txt
langchain
langchain-openai
langchain-community
langchain-pinecone
pinecone
pypdf
python-dotenv